my_tech_start/requirements.txt
aioredis==1.3.0
asgiref==3.2.2
asn1crypto==1.1.0
async-timeout==3.0.1
attrs==19.2.0
autobahn==19.10.1
Automat==0.7.0
backports.csv==1.0.7
certifi==2019.9.11
cffi==1.12.3
channels==2.3.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cycler==0.10.0
daphne==2.3.0
defusedxml==0.6.0
diff-match-patch==20181111
dill==0.3.1.1
dj-database-url==0.5.0
Django==2.2.5
django-common-helpers==0.9.2
django-compat==1.0.15
django-crontab==0.7.1
django-extensions==2.2.3
django-heroku==0.3.1
django-redis==4.10.0
djangorestframework==3.10.3
et-xmlfile==1.0.1
geographiclib==1.50
geopy==1.20.0
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4.1
jieba==0.39
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.1
msgpack==0.6.2
msgpack-python==0.5.6
numpy==1.17.2
odfpy==1.4.0
openpyxl==3.0.0
pandas==0.25.1
pbr==3.1.1
Pillow==6.1.0
psycopg2==2.8.3
pycparser==2.19
pycrypto==2.6.1
PyHamcrest==1.9.0
pyparsing==2.4.2
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
PyYAML==5.1.2
redis==2.10.6
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
South==1.0.2
sqlparse==0.3.0
tablib==0.13.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.6
whitenoise==4.1.4
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
